Fix starting index to 1.
We need to always be able to revert to previous index of 0 and get
a term of 0.  If we actually start using the leader's log index @0
then the previous term will always be 1 or 2.  In order to start to
have the leader's previous term be 0 and the followers' previous term
also be 0 we must start at log index 1 or handle a special case where
we check for -1 all the time.